Move backups to install file and restore depmod
The backups were a chicken & egg problem. If the install hasn't run on a
new system, the backup script isn't there yet. So instead of using a
script, just make it part of the pre_install.

Somehwere along the line we lost depmod being run. Add this back into
the initramfs hook, and since the initramfs hook runs post-transaction,
we can rely on a shell script.

Signed-off-by: Stephano Cetola <stephanoc@gmail.com>